Transfer the prize submission view to its own file
Same issue as the donate name clash. Also, I plan to have this file expand
in scope as more work gets done.
